move: 패키지 구조 변경
category 내부에 component 패키지 생성해서 관련 데이터를 옮김.
