+++ src/main/java/com/dominest/dominestbackend/api/category/request/CategoryCreateRequest.java
@@ -1,6 +1,6 @@
 package com.dominest.dominestbackend.api.category.request;
 
-import com.dominest.dominestbackend.domain.post.component.category.component.categorytype.Type;
+import com.dominest.dominestbackend.domain.post.component.category.component.Type;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import org.hibernate.validator.constraints.Length;
