+++ src/main/java/com/dominest/dominestbackend/api/category/response/CategoryListDto.java
@@ -1,7 +1,7 @@
 package com.dominest.dominestbackend.api.category.response;
 
 import com.dominest.dominestbackend.domain.post.component.category.Category;
-import com.dominest.dominestbackend.domain.post.component.category.component.categorytype.Type;
+import com.dominest.dominestbackend.domain.post.component.category.component.Type;
 import lombok.Builder;
 import lombok.Getter;
