+++ src/main/java/com/dominest/dominestbackend/domain/post/component/category/component/Type.java
@@ -1,0 +1,8 @@
+package com.dominest.dominestbackend.domain.post.component.category.component;
+
+import lombok.Getter;
+
+@Getter
+public enum Type {
+    IMAGE, TEXT_AND_IMAGE
+}
