--- src/main/java/com/dominest/dominestbackend/domain/post/component/category/component/categorytype/Type.java
@@ -1,8 +1,0 @@
-package com.dominest.dominestbackend.domain.post.component.category.component.categorytype;
-
-import lombok.Getter;
-
-@Getter
-public enum Type {
-    IMAGE, TEXT_AND_IMAGE
-}
